refactor: 분리한 훅을 사용하는 코드로 변경, 폴더구조 반영에 따른 수정사항
