+++ frontend/src/@components/@shared/Header/index.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 
 import Icon from '@/@components/@shared/Icon';
-import useMe from '@/hooks/useMe';
+import useMe from '@/@hooks/user/useMe';
 import { PATH } from '@/Router';
 import theme from '@/styles/theme';
